Move interpreter utils to their proper module
